+++ changelog.md
@@ -95,6 +95,9 @@ proc enumToString*(enums: openArray[enum]): string =
 
 - Added `os.relativePath`.
 - Added `parseopt.remainingArgs`.
+- Added `os.staticGetAppFilename`, which is `getAppFilename` at CT;
+  can be used to retrive the currently executing nim compiler.
+
 
 ### Library changes